Refactor: Move index creation to a new indexer module
The index.rs file was a bit large, so this splits up index creation and
index querying.

Also, it was a bit sketchy that to create an empty StrobemerIndex, one had
to do `index = StrobemerIndex::new()` followed by `index.populate()`, so in
between those steps, you had a half-initialized index. Now, there is only a
`make_index()` function that returns a fully populated index.
