+++ src/main.rs
@@ -14,13 +14,14 @@ use clap::builder::styling::AnsiColor;
 use fastrand::Rng;
 use log::{debug, error, info, trace, warn};
 use mimalloc::MiMalloc;
+use strobealign::indexer::make_index;
 use thiserror::Error;
 
 use strobealign::aligner::{Aligner, Scores};
 use strobealign::chainer::{Chainer, ChainingParameters};
 use strobealign::details::Details;
 use strobealign::index::{
-    IndexReadingError, REF_RANDSTROBE_MAX_NUMBER_OF_REFERENCES, StrobemerIndex,
+    IndexReadingError, REF_RANDSTROBE_MAX_NUMBER_OF_REFERENCES, StrobemerIndex, read_index,
 };
 use strobealign::insertsize::InsertSizeDistribution;
 use strobealign::io::SequenceIOError;
@@ -445,23 +446,24 @@ fn run() -> Result<(), CliError> {
     }
 
     // Create the index
-    let mut index = StrobemerIndex::new(&references, parameters.clone(), args.bits);
     debug!("Auxiliary hash length: {}", args.aux_len);
     info!("Multi-context seed strategy: {}", args.mcs_strategy);
-    info!("Bits used to index buckets: {}", index.bits);
+    //info!("Bits used to index buckets: {}", index.bits);
 
-    if args.use_index {
+    let index = if args.use_index {
         // Read the index from a file
         assert!(!args.create_index);
         let read_index_timer = Instant::now();
         let mut sti_path = args.ref_path.clone();
         sti_path.push_str(&parameters.filename_extension());
         info!("Reading index from {}", sti_path);
-        index.read(&sti_path)?;
+        let index = read_index(&sti_path, &references, parameters.clone(), args.bits)?;
         info!(
             "Total time reading index: {:.2} s",
             read_index_timer.elapsed().as_secs_f64()
         );
+
+        index
     } else {
         let timer = Instant::now();
         let indexing_threads = args.indexing_threads.unwrap_or(args.threads);
@@ -470,14 +472,21 @@ fn run() -> Result<(), CliError> {
             indexing_threads,
             if indexing_threads == 1 { "" } else { "s" }
         );
-        let index_stats = index.populate(args.filter_fraction, indexing_threads);
+        let (index, index_stats) = make_index(
+            &references,
+            parameters.clone(),
+            args.bits,
+            args.filter_fraction,
+            indexing_threads,
+        );
         info!(
             "Total time indexing: {:.2} s",
             timer.elapsed().as_secs_f64()
         );
         debug!("{}", &index_stats);
-    }
-    let index = index;
+
+        index
+    };
     debug!("Filtered cutoff count: {}", index.filter_cutoff);
     debug!("Using rescue cutoff: {}", index.rescue_cutoff);
